+++ packages/web/tests/stream.test.ts
@@ -1,101 +1,9 @@
-import { Schema, Table, column } from '@powersync/common';
-import { WebPowerSyncOpenFactoryOptions } from '@powersync/web';
 import Logger from 'js-logger';
-import { v4 as uuid } from 'uuid';
 import { afterEach, beforeAll, describe, expect, it, vi } from 'vitest';
-import { MockRemote, MockStreamOpenFactory, TestConnector } from './utils/MockStreamOpenFactory';
+import { TestConnector } from './utils/MockStreamOpenFactory';
+import { ConnectedDatabaseUtils, generateConnectedDatabase } from './utils/generateConnectedDatabase';
 
-type UnwrapPromise<T> = T extends Promise<infer U> ? U : T;
-
-export type ConnectedDatabaseUtils = UnwrapPromise<ReturnType<typeof generateConnectedDatabase>>;
-export type GenerateConnectedDatabaseOptions = {
-  powerSyncOptions: Partial<WebPowerSyncOpenFactoryOptions>;
-};
-
-// const UPLOAD_TIMEOUT_MS = 3000;
-const UPLOAD_TIMEOUT_MS = 30_000;
-
-export const DEFAULT_CONNECTED_POWERSYNC_OPTIONS = {
-  powerSyncOptions: {
-    dbFilename: 'test-stream-connection.db',
-    flags: {
-      enableMultiTabs: false,
-      useWebWorker: true
-    },
-    // Makes tests faster
-    crudUploadThrottleMs: 0,
-    schema: new Schema({
-      users: new Table({ name: column.text })
-    })
-  }
-};
-
-export async function generateConnectedDatabase(
-  options: GenerateConnectedDatabaseOptions = DEFAULT_CONNECTED_POWERSYNC_OPTIONS
-) {
-  const { powerSyncOptions } = options;
-  const { powerSyncOptions: defaultPowerSyncOptions } = DEFAULT_CONNECTED_POWERSYNC_OPTIONS;
-  /**
-   * Very basic implementation of a listener pattern.
-   * Required since we cannot extend multiple classes.
-   */
-  const callbacks: Map<string, () => void> = new Map();
-  const connector = new TestConnector();
-  const uploadSpy = vi.spyOn(connector, 'uploadData');
-  const remote = new MockRemote(connector, () => callbacks.forEach((c) => c()));
-
-  const factory = new MockStreamOpenFactory(
-    {
-      ...defaultPowerSyncOptions,
-      ...powerSyncOptions,
-      flags: {
-        ...(defaultPowerSyncOptions.flags ?? {}),
-        ...(powerSyncOptions.flags ?? {})
-      }
-    },
-    remote
-  );
-  const powersync = factory.getInstance();
-
-  const waitForStream = () =>
-    new Promise<void>((resolve) => {
-      const id = uuid();
-      callbacks.set(id, () => {
-        resolve();
-        callbacks.delete(id);
-      });
-    });
-
-  const connect = async () => {
-    const streamOpened = waitForStream();
-
-    const connectedPromise = powersync.connect(connector);
-
-    await streamOpened;
-
-    remote.streamController?.enqueue(new TextEncoder().encode('{"token_expires_in":3426}\n'));
-
-    // Wait for connected to be true
-    await connectedPromise;
-  };
-
-  await connect();
-
-  return {
-    connector,
-    connect,
-    factory,
-    powersync,
-    remote,
-    uploadSpy,
-    waitForStream,
-    dispose: async () => {
-      remote.streamController?.close();
-      await powersync.disconnectAndClear();
-      await powersync.close();
-    }
-  };
-}
+const UPLOAD_TIMEOUT_MS = 3000;
 
 describe(
   'Streaming',
